mockito-1 - Renaming artifacts and package name
