+++ junit-examples/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>junit-examples</name>
+	<name>unit-testing-examples</name>
 	<comment></comment>
 	<projects>
 	</projects>
