+++ junit-examples/src/main/java/com/hargun/unittesting/junit/StringHelper.java
@@ -1,4 +1,4 @@
-package com.hargun.junit.lecture1;
+package com.hargun.unittesting.junit;
 
 /**
  * @author hargun.suri
