+++ junit-examples/src/test/java/com/hargun/junit/lecture1/StringHelperAssertBasicExampleTest.java
@@ -4,6 +4,8 @@ import static org.junit.Assert.*;
 
 import org.junit.Test;
 
+import com.hargun.unittesting.junit.StringHelper;
+
 /**
  * @author hargun.suri
  *
