+++ junit-examples/src/test/java/com/hargun/junit/lecture1/StringHelperBeforeAftertBasicExampleTest.java
@@ -9,6 +9,8 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+import com.hargun.unittesting.junit.StringHelper;
+
 /**
  * @author hargun.suri
  *
